refactor: io 상수 별도관리하게 변경
