+++ src/main/java/racingcar/io/OutputManager.java
@@ -1,12 +1,14 @@
-package racingcar;
+package racingcar.io;
+
+import static racingcar.io.ErrorMessage.GUIDE_MESSAGE;
+import static racingcar.io.ErrorMessage.TRY_COUNT_MESSAGE;
+import static racingcar.io.ErrorMessage.WINNER_RESULT_MESSAGE;
 
 import java.util.List;
 
 public class OutputManager {
 
-    private static final String GUIDE_MESSAGE = "경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)";
-    private static final String WINNER_RESULT_MESSAGE = "최종 우승자 : ";
-    private static final String TRY_COUNT_MESSAGE = "시도할 횟수는 몇 회인가요?";
+
     private static final String DELIMITER = ", ";
 
     public static void printGuideMessage() {
