+++ src/test/java/racingcar/io/OutputManagerTest.java
@@ -1,7 +1,6 @@
-package racingcar;
+package racingcar.io;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 import java.io.ByteArrayOutputStream;
 import java.io.OutputStream;
